Cleaned up code
- MCSceneRenderer takes now care of the Minecraft Scene. It is loaded
from mainwindow.
- MCSceneRenderer inherits from World and has its own build function.
- cameras in the Minecraft scene are now concrete classes (not only the
abstract base class), the ambient light is now an AmbientOccluder.

- communication between the rendering camera and the mainwindow /
ImageDisplay through signals/slots are cleaner, could be further
improved. For example by generalizing the display_line signal to future
compatibility with OpenGL / GLUT.

- GUI improvements:
  > The focal length and focus distance of the camera are now set by
angle and distance. Angle is the perspective angle (horizontally) and
distance is the distance of the plane in focus.
 > downsampling was abandonned.
 > the mainwindow communicates with the world thread with the help of
render() and stoprender() to ensure that the camera doesn't try to
access samplers or viewplanes that are being reset.